Rest API

+++ restapi/src/main/java/org/fnovella/project/grade/controller/GradeController.java
@@ -1,12 +1,16 @@
 package org.fnovella.project.grade.controller;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import org.fnovella.project.course.repository.CourseRepository;
 import org.fnovella.project.grade.model.Grade;
 import org.fnovella.project.grade.repository.GradeRepository;
+import org.fnovella.project.inscriptions_inst_grade.repository.InscriptionsInstGradeRepository;
+import org.fnovella.project.inscriptions_part_grade.repository.InscriptionsPartGradeRepository;
 import org.fnovella.project.utility.model.APIResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -22,10 +26,14 @@ public class GradeController {
 	private GradeRepository gradeRepository;
 	@Autowired
 	private CourseRepository courseRepository;
+	@Autowired
+	private InscriptionsInstGradeRepository inscriptionsInstGradeRepository;
+	@Autowired
+	private InscriptionsPartGradeRepository inscriptionsPartGradeRepository;
 	
 	@RequestMapping(value = "", method = RequestMethod.GET)
-	public APIResponse getAll(@RequestHeader("authorization") String authorization) {
-		return new APIResponse(this.gradeRepository.findAll(), null);
+	public APIResponse getAll(@RequestHeader("authorization") String authorization, Pageable pageable) {
+		return new APIResponse(this.gradeRepository.findAll(pageable), null);
 	}
 	
 	@RequestMapping(value = "{gradeId}/courses", method = RequestMethod.GET)
@@ -59,7 +67,8 @@ public class GradeController {
 		ArrayList<String> errors = new ArrayList<String>();
 		Grade toUpdate = this.gradeRepository.findOne(id);
 		if (toUpdate != null) {
-			return new APIResponse(this.gradeRepository.saveAndFlush(grade), null);
+			toUpdate.setUpdateFields(grade);
+			return new APIResponse(this.gradeRepository.saveAndFlush(toUpdate), null);
 		}
 		errors.add("Grade doesn't exist");
 		return new APIResponse(null, errors);
@@ -70,7 +79,15 @@ public class GradeController {
 		ArrayList<String> errors = new ArrayList<String>();
 		Grade toDelete = this.gradeRepository.findOne(id);
 		if (toDelete != null) {
-			this.courseRepository.deleteByGrade(toDelete.getId());
+			List<?> list = this.inscriptionsPartGradeRepository.findByGradeId(id);
+			if (list != null && !list.isEmpty())
+				this.inscriptionsPartGradeRepository.deleteByGradeId(id);
+			list = this.inscriptionsInstGradeRepository.findByGradeId(id);
+			if (list != null && !list.isEmpty())
+				this.inscriptionsInstGradeRepository.deleteByGradeId(id);
+			list = this.courseRepository.findByGrade(id);
+			if (list != null && !list.isEmpty())
+				this.courseRepository.deleteByGrade(toDelete.getId());
 			this.gradeRepository.delete(toDelete);
 			return new APIResponse(true, null);
 		}
